MAR-3025. Fix tests.
